1st-July/requirements.txt
Flask
python-dotenv
requests
gTTS